added user role model and base model
