+++ models/__init__.py
@@ -1,2 +1,4 @@
+from models.base import BaseModel
+from models.user_role import UserRoleModel
 from models.user import UserModel
 from models.vehicle import VehicleModel
