+++ models/user.py
@@ -1,37 +1,26 @@
-import uuid
 from passlib.hash import pbkdf2_sha256
 
 from db import db
+from models import BaseModel
 
 
-class UserModel(db.Model):
+class UserModel(BaseModel):
     __tablename__ = "users"
 
-    
-    id = db.Column(db.String(32), primary_key=True)
     username = db.Column(db.String(80), nullable=False)
     email = db.Column(db.String(254), unique=True, nullable=False)
     password = db.Column(db.String(256), nullable=False)
-    role = db.Column(db.Integer, nullable=False)
-    vehicles = db.relationship("VehicleModel", back_populates="user", lazy="dynamic")
+    role_id = db.Column(db.String(32), db.ForeignKey(
+        "roles.id"), unique=False, nullable=False)
+
+    role = db.relationship("UserRoleModel")
+    vehicles = db.relationship(
+        "VehicleModel", back_populates="user", lazy="dynamic")
 
     def __init__(self, **kwargs):
         super(UserModel, self).__init__(**kwargs)
-        self.id = uuid.uuid4().hex
         self.password = pbkdf2_sha256.hash(kwargs["password"])
 
     @classmethod
-    def find_by_id(cls, id):
-        return cls.query.filter_by(id=id).first()
-    
-    @classmethod
     def find_by_email(cls, email):
         return cls.query.filter_by(email=email).first()
-    
-    def save_to_db(self):
-        db.session.add(self)
-        db.session.commit()
-
-    def delete_from_db(self):
-        db.session.delete(self)
-        db.session.commit()
